requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart
httpx

# ── Data ──────────────────────────────────────────────────────────────────────
pandas
numpy<2.0.0
scikit-learn
scipy

# ── ML Models ─────────────────────────────────────────────────────────────────
xgboost
lightgbm
prophet
statsmodels
joblib

# ── MLflow ────────────────────────────────────────────────────────────────────
mlflow

# ── Prefect ───────────────────────────────────────────────────────────────────
prefect

# ── Testing ───────────────────────────────────────────────────────────────────
pytest
pytest-asyncio
deepchecks

# ── Code Quality ──────────────────────────────────────────────────────────────
ruff



# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv
pydantic
pydantic-settings
requests
aiofiles
